requirements.txt
streamlit
sqlmodel
psycopg2-binary
requests
python-dotenv
scipy
pandas
sqlalchemy
rapidfuzz
